+++ web/actions/worker.py
@@ -119,7 +119,7 @@ def generate_warnings(worker):
             _("Please use a consistent Machinaris version to avoid issues."), 'warning'))
     if worker.fullnode_db_version() == "v1":  # None for non-fullnode, else 'v2' for already updated.
         # Don't warn if the blockchain is really old and supports only v1, so can't upgrade
-        if not worker.blockchain in ['flora','hddcoin','maize','mmx','nchain','silicoin','stor']:
+        if not globals.legacy_blockchain(worker.blockchain):
             wiki_link = "https://github.com/guydavis/machinaris/wiki/Forks#database-upgrade"
             if worker.blockchain == "chia":
                 wiki_link = "https://github.com/guydavis/machinaris/wiki/Chia#database-upgrade"